requirements.txt
Flask==3.0.1
Flask-Cors==4.0.1
Flask-Mail==0.10.0
Flask-MySQL==1.5.2
python-dotenv==1.0.1

